+++ CHANGELOG.md
@@ -8,6 +8,7 @@ Nedenfor ses dato for release og beskrivelse af opgaver som er implementeret.
 
 * Ændrede kort date/time format
 * Ændrede PDF udseende [PR-377](https://github.com/itk-dev/os2forms_selvbetjening/pull/377).
+* Tilføjede os2forms_custom_view_builders module [PR-379](https://github.com/itk-dev/os2forms_selvbetjening/pull/379).
 
 ## [3.2.4] 2025-01-07